common: Clean up filesystem.cpp
Split into files for all OSes, Windows, Linux/macOS, Linux and macOS.
